😊😊😊移动常量位置
